pyproject.toml
[project]
name = "xhs-fashion"
version = "0.1.0"
description = "XiaoHongShu fashion search + AI virtual try-on, with Claude Code / OpenClaw skill"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "typer>=0.12,<1.0",
    "Pillow>=11.0.0",
]

[project.scripts]
xhs = "xhs_cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["xhs_cli"]
